backend/app/models/_objectid.py
"""
Custom Pydantic-compatible ``ObjectId`` type for MongoDB documents.

This lets Pydantic v2 serialise / deserialise BSON ObjectIds transparently
so that API responses contain plain hex strings while internal code can
work with native ``ObjectId`` instances.
"""

from typing import Annotated, Any

from bson import ObjectId
from pydantic import GetCoreSchemaHandler, GetJsonSchemaHandler
from pydantic.json_schema import JsonSchemaValue
from pydantic_core import CoreSchema, core_schema


class _ObjectIdValidator:
    """Pydantic v2 custom type for ``bson.ObjectId``."""

    @classmethod
    def __get_pydantic_core_schema__(
        cls,
        _source_type: Any,
        _handler: GetCoreSchemaHandler,
    ) -> CoreSchema:
        return core_schema.no_info_plain_validator_function(
            cls._validate,
            serialization=core_schema.to_string_ser_schema(),
        )

    @classmethod
    def __get_pydantic_json_schema__(
        cls,
        _schema: CoreSchema,
        handler: GetJsonSchemaHandler,
    ) -> JsonSchemaValue:
        return handler(core_schema.str_schema())

    @classmethod
    def _validate(cls, value: Any) -> ObjectId:
        if isinstance(value, ObjectId):
            return value
        if isinstance(value, str) and ObjectId.is_valid(value):
            return ObjectId(value)
        raise ValueError(f"Invalid ObjectId: {value!r}")


PyObjectId = Annotated[ObjectId, _ObjectIdValidator]
