backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
motor>=3.6,<4
pymongo>=4.9,<5
pydantic>=2.10,<3
pydantic-settings>=2.7,<3
google-genai>=1.0,<2
elevenlabs>=1.50,<2
python-dotenv>=1.1,<2
httpx>=0.28,<1
